+++ CHANGELOG.md
@@ -1,14 +1,17 @@
 # CHANGELOG
 
-
 ### 1.0.0-beta.0
 
 - Use `event-emitter` as deps, no need addtional install
 - Change build: replace `tsdx` with `bunchee`
 - Bump example depencies
 
-
 ### 1.0.0-beta.1
 
 - Clean types
 - Use `pnpm` for install instead `yarn`
+
+### 1.0.0-beta.2
+
+- Clean types again
+- `syncState` and `replaceState` support `oldState => newState`